// src/infrastructure/interfaces/teslo-products.response.ts
// Generated by https://quicktype.io

export interface TesloProduct {
  title: string;
  id: number;
  price: number;
  description: string;
  slug: string;
  stock: number;
  sizes: any[];
  gender: Gender;
  tags: any[];
  images: string[];
  owner: Owner;
  predictions: Prediction[];
}

export enum Gender {
  Kid = "kid",
  Men = "men",
  Unisex = "unisex",
  Women = "women",
}

export enum Size {
  L = "L",
  M = "M",
  S = "S",
  Xl = "XL",
  Xs = "XS",
  Xxl = "XXL",
}


export interface TesloUser {
  id: string;
  email: string;
  fullName: string;
  isActive: boolean;
  roles: string[];
}



export interface Owner {
  email: string;
  is_active: boolean;
  is_superuser: boolean;
  full_name: null;
  id: number;
}

export interface Prediction {
  height: number;
  x: number;
  class_name: ClassName;
  detection_id: string;
  prediction_type: PredictionType;
  product_id: number;
  width: number;
  y: number;
  confidence: number;
  class_id: number;
  image_path: string;
  id: number;
}

export enum ClassName {
  Flor = "flor",
  Inmadura = "inmadura",
  Pintona = 'pintona',
  Madura = "madura",
  Podrida = 'podrida',
}


export enum PredictionType {
  ObjectDetectionModel = "ObjectDetectionModel",
}